Covid webapp
 It is a single page web application, which shows the number of covid cases
based on static data till july 2020. It shows the red circle on map, bigger the circle higher the
number of covid active cases. After hovering on any circle it pops out the active cases at that
place. Along with the map it also displays the top 15 countries covid active cases in the
panel. The technologies I have used are HTML CSS, python’s flask, folium and pandas.
Flask as a web framework. I have used data from csv files and to access required data in
required format I used Pandas library. To display the map I have used the folium library and
mark the red circle on it according to the number of test cases